backend/requirements.txt
Flask==3.0.2
Flask-PyMongo==2.3.0
Flask-Cors==4.0.0
pymongo==4.6.3
python-bsonjs==0.3.0
requests==2.31.0
pytz==2023.4